# docs/openapi-alerts.json
{
  "components": {
    "schemas": {
      "ActionabilityRequest": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Type"
          },
          "actionable": {
            "pattern": "^(YES|NO|AMBIGUOUS)$",
            "title": "Actionable",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "delivery_id": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Id"
          },
          "episode_id": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Episode Id",
            "type": "string"
          },
          "reason_code": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "required": [
          "episode_id",
          "actionable"
        ],
        "title": "ActionabilityRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Alert-system subset of the bubblegauge OpenAPI document, generated from the running application. Reads use ALERTS_READ_API_KEY, writes use ALERTS_WRITE_API_KEY; neither falls back to the admin key. Research, not advice.",
    "title": "bubblegauge \u2014 alerts",
    "version": "3.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/alerts/actionability": {
      "post": {
        "description": "One human label per alert, for the question the metrics cannot answer.\n\nEverything else in this system measures whether a message went out;\nnothing measures whether it was worth receiving. Stage 7 retains the LLM\nselector only if these labels show a material improvement, and AMBIGUOUS\nexists so an unsure reviewer does not inflate the KPI either way.",
        "operationId": "record_actionability_api_v1_admin_alerts_actionability_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionabilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Record Actionability Api V1 Admin Alerts Actionability Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record whether an alert was actionable",
        "tags": [
          "alerts-admin"
        ]
      }
    },
...
    "/api/v1/admin/alerts/evaluate": {
      "post": {
        "description": "Run P0b/P1/P2 for a single sidecar without waiting for a recompute.\n\n`shadow=true` (the default) evaluates into the shadow namespace, which is\nhow a promoted ruleset is exercised before anything is allowed to send.",
        "operationId": "admin_evaluate_api_v1_admin_alerts_evaluate_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Admin Evaluate Api V1 Admin Alerts Evaluate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate one captured input",
        "tags": [
          "alerts-admin"
        ]
      }
    },
    "/api/v1/admin/alerts/promote": {
      "post": {
        "description": "Validate the artifacts on disk and PROMOTE them.\n\nDeliberately an explicit action with its own endpoint: nothing promotes as\na side effect of a boot, a deploy or a validation run. Promotion does not\nchange ALERTS_MODE \u2014 going live is still a separate operator decision.",
        "operationId": "admin_promote_api_v1_admin_alerts_promote_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Admin Promote Api V1 Admin Alerts Promote Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote the candidate ruleset",
        "tags": [
          "alerts-admin"
        ]
      }
    },
    "/api/v1/admin/alerts/recover": {
      "post": {
        "operationId": "admin_recover_api_v1_admin_alerts_recover_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Admin Recover Api V1 Admin Alerts Recover Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sweep stale evaluation leases",
        "tags": [
          "alerts-admin"
        ]
      }
    },
    "/api/v1/admin/alerts/renders/{render_id}": {
      "get": {
        "description": "The operator path to a rendered message body.\n\nThe read surface withholds `final_message` because the frontend uses a\nbrowser-visible scoped token (H-05), and the alert scopes deliberately do\nnot nest \u2014 there is no stronger key a caller could present to\n`/api/v1/alerts/renders/{id}`. So the text lives here, behind admin, and\n`no-store` keeps it out of any intermediary cache.",
        "operationId": "admin_get_render_api_v1_admin_alerts_renders__render_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "render_id",
            "required": true,
            "schema": {
              "title": "Render Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Admin Get Render Api V1 Admin Alerts Renders  Render Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One render, INCLUDING the text",
        "tags": [
          "alerts-admin"
        ]
      }
    },
    "/api/v1/admin/alerts/send-test": {
      "post": {
        "description": "Create a TEST delivery for the dispatcher to send.\n\nTEST is the one delivery kind allowed zero members: it is about the\nTRANSPORT, not about any market condition, and inventing an episode to\nhang it on would put a fake market event in the audit trail. It is also\noutside `BUDGETED_KINDS`, so proving the wire works never spends the\noperator's non-P1 budget \u2014 and its body is a reviewed phrase-set fragment\nlike every other message, not prose typed into a request.\n\nThe actual send happens through the ordinary dispatcher: same claim, same\nadmission, same classification. A test that bypassed the pipeline would\nprove something other than the thing the operator needs proven.",
        "operationId": "send_test_api_v1_admin_alerts_send_test_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Send Test Api V1 Admin Alerts Send Test Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue an audited TEST delivery",
        "tags": [
          "alerts-admin"
        ]
      }
    },
...
  }
}